+++ backend/db/src/test/java/com/bls/patronage/dao/DeckDAOTest.java
@@ -1,10 +1,13 @@
 package com.bls.patronage.dao;
 
 import com.bls.patronage.db.dao.DeckDAO;
+import com.bls.patronage.db.exception.DataAccessException;
 import com.bls.patronage.db.model.Deck;
+import com.bls.patronage.db.model.DeckWithFlashcardsNumber;
 import org.junit.Test;
 import org.skife.jdbi.v2.Handle;
 
+import java.util.Collection;
 import java.util.List;
 import java.util.UUID;
 
@@ -19,66 +22,93 @@ public class DeckDAOTest extends DAOTest {
     public void setUp() throws Exception {
         super.setUp();
 
-        deckExample1 = new Deck(UUID.fromString("020276e9-285c-4162-b585-32f272a947b1"),"math");
-        deckExample2 = new Deck(UUID.fromString("b5e9aa75-64d6-4d21-87a6-d0a91c70f997"),"bio");
+        deckExample1 = new Deck(UUID.fromString("020276e9-285c-4162-b585-32f272a947b1"), "math", false);
+        deckExample2 = new Deck(UUID.fromString("b5e9aa75-64d6-4d21-87a6-d0a91c70f997"), "bio", true);
     }
 
     @Override
     protected void setUpDatabaseContent(Handle handle) {
         handle.createCall("DROP TABLE decks IF EXISTS").invoke();
         handle.createCall(
-                "CREATE TABLE decks (id uuid primary key, name varchar(50) not null)")
+                "CREATE TABLE decks (id uuid primary key, name varchar(50) not null, public boolean)")
                 .invoke();
-        handle.createStatement("INSERT INTO decks VALUES (?, ?)")
+        handle.createStatement("INSERT INTO decks VALUES (?, ?, ?)")
                 .bind(0, "020276e9-285c-4162-b585-32f272a947b1")
                 .bind(1, "math")
+                .bind(2, false)
                 .execute();
-        handle.createStatement("INSERT INTO decks VALUES (?, ?)")
+        handle.createStatement("INSERT INTO decks VALUES (?, ?, ?)")
                 .bind(0, "b5e9aa75-64d6-4d21-87a6-d0a91c70f997")
                 .bind(1, "bio")
+                .bind(2, true)
+                .execute();
+        handle.createCall("DROP TABLE flashcards IF EXISTS").invoke();
+        handle.createCall(
+                "CREATE TABLE flashcards (id uuid primary key, question varchar(1000) not null," +
+                        " answer varchar(1000) not null, deckid uuid not null)")
+                .invoke();
+        handle.createStatement("INSERT INTO flashcards VALUES (?, ?, ?, ?)")
+                .bind(0, "15bb54d4-bd08-40b4-a0fb-2f35865173fe")
+                .bind(1, "ping")
+                .bind(2, "pong")
+                .bind(3, "020276e9-285c-4162-b585-32f272a947b1")
                 .execute();
     }
     @Test
-    public void getAllDecks(){
+    public void getAllDecks() throws Exception {
         final DeckDAO dao = dbi.open(DeckDAO.class);
-        final List<Deck> decks = dao.getAllDecks();
-        assertThat(decks).containsSequence(deckExample1,deckExample2);
+        final Collection<Deck> decks = dao.getAllDecks();
+        assertThat(decks).containsSequence(deckExample1, deckExample2);
     }
 
     @Test
-    public void getDeckById(){
+    public void getAllDecksWithFlashcardsNumber() throws Exception {
+        final DeckWithFlashcardsNumber deckWithFlashcards1
+                = new DeckWithFlashcardsNumber(deckExample1.getId(),
+                "math", false, 1);
+        final DeckWithFlashcardsNumber deckWithFlashcards2
+                = new DeckWithFlashcardsNumber(deckExample2.getId(),
+                "bio", true, 0);
+
+        final DeckDAO dao = dbi.open(DeckDAO.class);
+        final Collection<DeckWithFlashcardsNumber> foundDecks = dao.getAllDecksWithFlashcardsNumber();
+        assertThat(foundDecks).contains(deckWithFlashcards1, deckWithFlashcards2);
+    }
+
+    @Test
+    public void getDeckById() throws Exception {
         final DeckDAO dao = dbi.open(DeckDAO.class);
         final Deck foundDeck = dao.getDeckById(deckExample1.getId());
         assertThat(foundDeck).isEqualTo(deckExample1);
     }
 
     @Test
-    public void getDeckByName(){
+    public void getDeckByName() throws Exception {
         final DeckDAO dao = dbi.open(DeckDAO.class);
-        final List<Deck> decks = dao.getDecksByName(deckExample1.getName());
+        final Collection<Deck> decks = dao.getDecksByName(deckExample1.getName());
         assertThat(decks).containsOnly(deckExample1);
     }
 
     @Test
-    public void createDeck(){
-        final Deck createdDeck = new Deck(UUID.fromString("a04692bc-4a70-4696-9815-24b8c0de5398"),"sport");
+    public void createDeck() throws Exception {
+        final Deck createdDeck = new Deck(UUID.fromString("a04692bc-4a70-4696-9815-24b8c0de5398"), "sport", true);
         final DeckDAO dao = dbi.open(DeckDAO.class);
         dao.createDeck(createdDeck);
         assertThat(dao.getDeckById(createdDeck.getId())).isEqualTo(createdDeck);
     }
 
-    @Test
-    public void deleteDeck(){
+    @Test(expected = DataAccessException.class)
+    public void deleteDeck() throws Exception {
         final DeckDAO dao = dbi.open(DeckDAO.class);
         dao.deleteDeck(deckExample2.getId());
         assertThat(dao.getDeckById(deckExample2.getId())).isNull();
     }
 
     @Test
-    public void updateDeck(){
-        final Deck updatedDeck = new Deck(deckExample1.getId(),"miscellanous");
+    public void updateDeck() throws Exception {
+        final Deck updatedDeck = new Deck(deckExample1.getId(), "miscellaneous", true);
         final DeckDAO dao = dbi.open(DeckDAO.class);
-        dao.updateDeck(updatedDeck);
+        dao.update(updatedDeck);
         assertThat(dao.getDeckById(deckExample1.getId())).isEqualTo(updatedDeck);
     }
 }
